Native build now executes and deploys successfully from Android Studio. Restructuring Java package
